tidy: Move arm types to expanders unless shared
