+++ internal/pkg/expanders/subscription.go
@@ -38,7 +38,7 @@ func (e *SubscriptionExpander) Expand(ctx context.Context, currentItem *TreeNode
 
 	//    \/ It's not the usual ... look out
 	if err == nil {
-		var rgResponse armclient.ResourceGroupResponse
+		var rgResponse ResourceGroupResponse
 		err = json.Unmarshal([]byte(data), &rgResponse)
 		if err != nil {
 			panic(err)
@@ -67,4 +67,16 @@ func (e *SubscriptionExpander) Expand(ctx context.Context, currentItem *TreeNode
 		SourceDescription: "Resource Group Request",
 		IsPrimaryResponse: true,
 	}
+}
+
+// ResourceGroupResponse ResourceGroup rest type
+type ResourceGroupResponse struct {
+	Groups []struct {
+		ID         string `json:"id"`
+		Name       string `json:"name"`
+		Location   string `json:"location"`
+		Properties struct {
+			ProvisioningState string `json:"provisioningState"`
+		} `json:"properties"`
+	} `json:"value"`
 }
